% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_returns.R
\name{aggregate_returns}
\alias{aggregate_returns}
\title{Aggregate Returns by characteristic}
\usage{
aggregate_returns(
  df_returns,
  df_info,
  aggregate.by = mkt_sector,
  col.id = id,
  col.weight = weight,
  col.date = date,
  col.ret = ret
)
}
\arguments{
\item{df_returns}{Return data frame, usually generated by \code{security_returns()}}

\item{df_info}{Info data frame, downloaded by \code{download_port_bbg()}}

\item{aggregate.by}{Column of info data frame to aggregate the returns by}
}
\value{
Return aggregated return data frame
}
\description{
Aggregate Returns by characteristic
}
