% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_demo_data.R
\docType{data}
\name{bl_data}
\alias{bl_data}
\title{Sample Return data for the Black Litterman Model}
\format{
Data frame with three columns
\describe{
\item{date}{Date}
\item{id}{ID of the asset class}
\item{px}{The (weekly) return}
}
}
\usage{
bl_data
}
\description{
This data set can be used to run the demo for the black litterman model.
}
\examples{
# load the data
data(bl_data)

# run the demo
demo("bl_demo")


}
\keyword{datasets}
