% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_price_NAs.R
\name{fill_price_NAs}
\alias{fill_price_NAs}
\title{Fill Missing Values in Price Time-Series}
\usage{
fill_price_NAs(df, col.date, col.id, col.px, method = "interpolate")
}
\arguments{
\item{df}{Data frame containing price data}

\item{col.date}{Name of the date column}

\item{col.id}{Name of the id column}

\item{col.px}{Name of the price column}

\item{method}{Method to handle missing values. Options: "interpolate", "forward", "backward"}
}
\value{
A data frame with missing values filled
}
\description{
This function fills missing values in a price time-series dataset using a chosen method.
}
