% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bond_data.R
\name{get_bond_data}
\alias{get_bond_data}
\title{Get Government Bond Data}
\usage{
get_bond_data()
}
\value{
A data frame containing government bond data.
}
\description{
This function scrapes government bond data from "worldgovernmentbonds.com", cleans the data, and returns a table containing data on 10Y Bond Yields, Spread vs Bund, Spread vs T-Note and Spread vs Bank Rate.
}
