% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_price_NAs.R
\name{plot_price_NAs}
\alias{plot_price_NAs}
\title{Visualize NAs in price time-series}
\usage{
plot_price_NAs(df, col.date = date, col.id = id, col.px = px)
}
\arguments{
\item{df}{Price data frame (usually downloaded via \code{pmp.engine::download_port_bbg})}

\item{col.date}{Name of the date column}

\item{col.id}{Name of the id column}

\item{col.px}{Name of the price column}
}
\value{
Returns a plot
}
\description{
Function to get an idea where NAs are in the data set. Recommened to take a look before running any analysis.
}
