% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_id.R
\docType{data}
\name{port_id}
\alias{port_id}
\title{Sample ID data}
\format{
Data frame with six columns
\describe{
\item{name}{Name of the position}
\item{issuer}{Issuer, mainly relevant for bonds}
\item{crncy}{Currency the position is denominated in}
\item{mkt_sector}{Market Sector (sometimes a bit arbitrary)}
\item{id}{The BBG id of the position}
\item{weight}{Weight of the position in our portfolio}
}
}
\usage{
port_id
}
\description{
Downloading the portfolio assets using \code{download_port_bbg()}, returns two data frames. One is \code{id},
which contains several characteristics for each position in our portfolio. Most importantly the \code{weight}
within our portfolio.
}
\details{
This sample data frame can be used to test functions that are based on the \code{download_port_bbg()} function
data.

\emph{It should not be used for current analysis, as it might not be up to date}
}
\examples{
data(port_id)


}
\keyword{datasets}
