% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_price.R
\docType{data}
\name{port_price}
\alias{port_price}
\title{Sample Price data}
\format{
Data frame with three columns
\describe{
\item{date}{Date}
\item{id}{BBG ID of the position}
\item{px}{The last price of the day of our position}
}
}
\usage{
port_price
}
\description{
Downloading the portfolio assets using \code{download_port_bbg()}, returns two data frames. One is \code{price},
which contains daily prices for each position in our portfolio.
}
\details{
This sample data frame can be used to test functions that are based on the \code{download_port_bbg()} function
data. It includes price data for 2023.

\emph{It should not be used for current analysis, as it might not be up to date}
}
\examples{
data(port_price)


}
\keyword{datasets}
