% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_reporting.R
\name{portfolio_reporting}
\alias{portfolio_reporting}
\title{Automated Portfolio Reporting}
\usage{
portfolio_reporting(df_value)
}
\arguments{
\item{df_value}{Xts object containing historic value of the portfolio.}
}
\value{
Returns nothing currently
}
\description{
This function should automate large parts of the reporting.
}
\details{
The dataframe for df_value can be downloaded from llb.at, by clicking
Wealth -> Performance -> ... -> Export. Make sure to adjust the timeframe and save
as a csv. Convert the dataframe into an xts before inputing it.
}
