% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_returns.R
\name{security_returns}
\alias{security_returns}
\title{Compute Returns}
\usage{
security_returns(
  df_price,
  period = "weeks",
  col.id = id,
  col.px = px,
  col.date = date
)
}
\arguments{
\item{df_price}{Time-series price data frame}

\item{period}{Frequency of returns, based on \code{xts::to.period()}}
}
\value{
Data frame with all security returns
}
\description{
This function computes the return for each security on the portfolio.
}
\details{
Frequency can be one of: "days", "weeks", "months", "quarters" and "years".
}
