% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500tickers.R
\docType{data}
\name{sp500tickers}
\alias{sp500tickers}
\title{Data that includes the tickers and the corresponding company names of the SP500.}
\format{
Data frame with two columns
\describe{
\item{Ticker}{SP500 Company ticker}
\item{Name}{SP500 Company Name}
}
}
\usage{
sp500tickers
}
\description{
Data that includes the tickers and the corresponding company names of the SP500.
}
\examples{
data("sp500tickers")


}
\keyword{datasets}
