% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StdDev_decomposition.R
\name{StdDev_decomposition}
\alias{StdDev_decomposition}
\title{Volatility Portfolio Decomposition}
\usage{
StdDev_decomposition(
  df_returns,
  df_info,
  clean = "boudt",
  col.id = "id",
  col.weight = "weight",
  col.date = "date",
  col.ret = "ret"
)
}
\arguments{
\item{df_returns}{Return data frame, usually generated by \code{security_returns()}}

\item{df_info}{Info data frame, downloaded by \code{download_port_bbg()}}

\item{clean}{Cleaning Method for returns to choose. Default parameter is "boudt". Details can be found in the help section of \code{PerformanceAnalytics::StdDev}}

\item{col.id}{Name of the "id" column, should be the same for df_returns and df_info}

\item{col.weight}{Name of the weight column in df_info}

\item{col.date}{Name of the date column in df_returns}

\item{col.ret}{Name of the return column in df_returns}
}
\value{
Returns the a list of the risk decomposition by volatility for all assets in the portfolio.
}
\description{
This function computes the attribution of each asset in the portfolio
to the total portfolio volatility
}
\details{
The column names paramters should be given as a string.
}
