% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_decomposition.R
\name{VaR_decomposition}
\alias{VaR_decomposition}
\title{Value-at-Risk Portfolio Decomposition}
\usage{
VaR_decomposition(
  df_returns,
  df_info,
  clean = "boudt",
  method = "modified",
  sign = 0.95,
  col.id = "id",
  col.weight = "weight",
  col.date = "date",
  col.ret = "ret"
)
}
\arguments{
\item{df_returns}{Return data frame, usually generated by \code{security_returns()}}

\item{df_info}{Info data frame, downloaded by \code{download_port_bbg()}}

\item{clean}{Cleaning Method for returns to choose. Default parameter is "boudt". Details can be found in the help section of \code{PerformanceAnalytics::StdDev}}

\item{method}{Choose method to compute the VaR. Default parameter is "modified". Detail can be found in the help section of \code{PerformanceAnalytics::StdDev}}

\item{sign}{Significance level for the VaR}

\item{col.id}{Name of the "id" column, should be the same for df_returns and df_info}

\item{col.weight}{Name of the weight column in df_info}

\item{col.date}{Name of the date column in df_returns}

\item{col.ret}{Name of the return column in df_returns}
}
\value{
Returns the a list of the risk decomposition by VaR for all assets in the portfolio.
}
\description{
This function computes the attribution of each asset in the portfolio
to the total portfolio value at risk
}
\details{
The column names paramters should be given as a string.
}
