% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_litterman.R
\name{black_litterman}
\alias{black_litterman}
\title{Black Litterman Model
This model can be used for asset allocation.}
\usage{
black_litterman(
  returns,
  weights,
  tau,
  views,
  freq,
  lambda.symb,
  col.date,
  col.symb,
  col.return
)
}
\arguments{
\item{returns}{A matrix of returns of the asset classes.}

\item{weights}{Strategic / Current Weights of the asset classes.}

\item{tau}{Uncertainty parameter}

\item{views}{Views data frame should be of specific form given in details.}

\item{freq}{Frequency of returns, given as an integer.}

\item{col.date}{Name of the date columns in returns.}

\item{col.symb}{Name of the symbol column in returns.}

\item{col.return}{Name of the return column in returns.}

\item{lambda.simb}{Which column to use for computing the risk aversion, given as \code{char}}
}
\value{
results
}
\description{
Black Litterman Model
This model can be used for asset allocation.
}
